PI2021Django/requirements.txt
git+git://github.com/LuisValentim1/PI_DBOT_Lib@main#egg=DBoT
Django
djangorestframework
django-cassandra-engine
cassandra-driver